feat: implement label filtering for participant sessions
